requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
